+++ dpdk/lib/librte_eal/include/rte_eal_paging.h
@@ -61,7 +61,7 @@ enum rte_map_flags {
 __rte_internal
 void *
 rte_mem_map(void *requested_addr, size_t size, int prot, int flags,
-	int fd, size_t offset);
+	int fd, uint64_t offset);
 
 /**
  * OS-independent implementation of POSIX munmap(3).
